+++ Braintree/Merchant.cs
@@ -19,6 +19,13 @@ namespace Braintree
             CountryName = merchantNode.GetString("country-name");
 
             Credentials = new OAuthCredentials(node.GetNode("credentials"));
+
+            var merchantAccountXmlNodes = merchantNode.GetList("merchant-accounts/merchant-account");
+            MerchantAccounts = new MerchantAccount[merchantAccountXmlNodes.Count];
+            for (int i = 0; i < merchantAccountXmlNodes.Count; i++)
+            {
+                MerchantAccounts[i] = new MerchantAccount(merchantAccountXmlNodes[i]);
+            }
         }
 
         public OAuthCredentials Credentials;
@@ -54,6 +61,11 @@ namespace Braintree
         }
 
         public string CountryName
+        {
+            get; set;
+        }
+
+        public MerchantAccount[] MerchantAccounts
         {
             get; set;
         }